dyn_*: Fix compiler warnings
DYN_* functions now return a LONG instead of an int

dyn_unix.c:59:10: warning: implicit conversion from 'LONG' (aka 'long') to 'int'
      changes value from 2148532244 to -2146435052 [-Wconstant-conversion]
                return SCARD_F_UNKNOWN_ERROR;
                ~~~~~~ ^~~~~~~~~~~~~~~~~~~~~
../src/PCSC/pcsclite.h:147:33: note: expanded from macro 'SCARD_F_UNKNOWN_ERROR'
                                         ^~~~~~~~~~~~~~~~
dyn_unix.c:77:10: warning: implicit conversion from 'LONG' (aka 'long') to 'int'
      changes value from 2148532244 to -2146435052 [-Wconstant-conversion]
                return SCARD_F_UNKNOWN_ERROR;
                ~~~~~~ ^~~~~~~~~~~~~~~~~~~~~
../src/PCSC/pcsclite.h:147:33: note: expanded from macro 'SCARD_F_UNKNOWN_ERROR'
                                         ^~~~~~~~~~~~~~~~
dyn_unix.c:105:8: warning: implicit conversion from 'LONG' (aka 'long') to 'int'
      changes value from 2148532244 to -2146435052 [-Wconstant-conversion]
                rv = SCARD_F_UNKNOWN_ERROR;
                   ~ ^~~~~~~~~~~~~~~~~~~~~
../src/PCSC/pcsclite.h:147:33: note: expanded from macro 'SCARD_F_UNKNOWN_ERROR'
                                         ^~~~~~~~~~~~~~~~

+++ src/dyn_generic.h
@@ -38,8 +38,8 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #ifndef __dyn_generic_h__
 #define __dyn_generic_h__
 
-	int DYN_LoadLibrary(void **, char *);
-	int DYN_CloseLibrary(void **);
-	int DYN_GetAddress(void *, /*@out@*/ void **, const char *, int);
+	LONG DYN_LoadLibrary(void **, char *);
+	LONG DYN_CloseLibrary(void **);
+	LONG DYN_GetAddress(void *, /*@out@*/ void **, const char *, int);
 
 #endif
